# internal/apidocs/openapi.yaml
openapi: 3.0.3
info:
  title: Xet Server API
  version: 0.8.0
  description: |
    HTTP APIs exposed by `xetd`: a wire-compatible Xet CAS server
    (`internal/casserver`, mounted at `/v1` and `/v2`), an optional Hub API
    shim (`internal/hubserver`, mounted on a separate port via `-hub-addr`)
    that lets the real `hf` CLI talk to this server end-to-end, and the Xet
    Data API (`internal/api`, mounted at `/v1` alongside the CAS protocol,
    on a disjoint set of literal paths) - a simple, non-wire-compatible
    chunk/dedup API for quick manual testing via the `xet` CLI.

    ## Authentication

    Every endpoint except telemetry/storage-stats (CAS) and `/v1/stats`
    (Xet Data API) is gated by a pluggable `auth.Authenticator`,
    defaulting to no enforcement at all (this server's pre-v0.8.0
    behavior). Enabled via `xetd -auth-token <secret>` (or
    `$XETD_AUTH_TOKEN`/`$HF_TOKEN`), which installs a single shared bearer
    token checked against every request:

    - **401 Unauthorized** - no `Authorization` header, or a token that
      fails validation entirely (wrong token, malformed header).
    - **403 Forbidden** - a valid token that lacks the scope (`read` or
      `write`) the endpoint requires.

    In this UI, use **Authorize** and paste the same secret the server was
    started with; it is sent as `Authorization: Bearer <token>` on every
    "Try it out" call. If the server has no auth configured you can skip
...

    ## Trying endpoints from this page

...

...

...

...
servers:
...
security:
  - bearerAuth: []
components:
  responses:
    Unauthorized:
      description: No credential, or one that fails validation entirely.
      content:
        text/plain:
          schema: { type: string }
    Forbidden:
      description: Valid credential, but lacking the scope this endpoint requires.
      content:
        text/plain:
          schema: { type: string }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Optional. Only enforced if xetd was started with -auth-token (or
        $XETD_AUTH_TOKEN/$HF_TOKEN) set to something other than "None". A
        server with no auth configured accepts every request regardless of
        whether this header is present.
  schemas:
...
    IndexRange:
      type: object
      description: Chunk-index range; End is exclusive.
      properties:
        start: { type: integer, format: int64 }
        end: { type: integer, format: int64 }
      required: [start, end]
    ByteRange:
      type: object
      description: Byte range; End is inclusive.
      properties:
        start: { type: integer, format: int64 }
        end: { type: integer, format: int64 }
      required: [start, end]
    UploadXorbResponse:
      type: object
      properties:
        was_inserted:
          type: boolean
          description: false if this xorb hash was already stored (dedup hit).
      required: [was_inserted]
    UploadShardResponse:
      type: object
      properties:
        result:
          type: integer
          description: >-
            0 = already exists, 1 = SyncPerformed. This server always
            returns 1 (it has no separate shard-dedup store).
      required: [result]
    ReconstructionTerm:
      type: object
      properties:
        hash: { type: string, description: Xorb hash (hex) this term reads from. }
        range: { $ref: '#/components/schemas/IndexRange' }
        unpacked_length: { type: integer, format: int64 }
      required: [hash, range, unpacked_length]
    FetchInfoEntry:
      type: object
      properties:
        url: { type: string, format: uri }
        url_range: { $ref: '#/components/schemas/ByteRange' }
        range: { $ref: '#/components/schemas/IndexRange' }
      required: [url, url_range, range]
    ReconstructionResponseV1:
      type: object
      properties:
        offset_into_first_range: { type: integer, format: int64 }
        terms:
          type: array
          items: { $ref: '#/components/schemas/ReconstructionTerm' }
        fetch_info:
          type: object
          description: Keyed by xorb hash (hex).
          additionalProperties:
            type: array
            items: { $ref: '#/components/schemas/FetchInfoEntry' }
      required: [offset_into_first_range, terms, fetch_info]
    XorbRangeDescriptor:
      type: object
      properties:
        chunks: { $ref: '#/components/schemas/IndexRange' }
        bytes: { $ref: '#/components/schemas/ByteRange' }
      required: [chunks, bytes]
    XorbMultiRangeFetch:
      type: object
      properties:
        url: { type: string, format: uri }
        ranges:
          type: array
          items: { $ref: '#/components/schemas/XorbRangeDescriptor' }
      required: [url, ranges]
    ReconstructionResponseV2:
      type: object
      properties:
        offset_into_first_range: { type: integer, format: int64 }
        terms:
          type: array
          items: { $ref: '#/components/schemas/ReconstructionTerm' }
        xorbs:
          type: object
          description: Keyed by xorb hash (hex).
          additionalProperties:
            type: array
            items: { $ref: '#/components/schemas/XorbMultiRangeFetch' }
      required: [offset_into_first_range, terms, xorbs]
...
paths:
  /v1/xorbs/{prefix}/{hash}:
    post:
      tags: [CAS]
      summary: Upload a serialized xorb
      description: >-
        Requires write scope. {prefix} must be the literal "default".
        Body is opaque binary (chunk headers + compressed payloads, no
        footer - see internal/xorbformat), capped at 128 MiB.
      parameters:
        - { name: prefix, in: path, required: true, schema: { type: string, enum: [default] } }
        - { name: hash, in: path, required: true, schema: { type: string }, description: Merkle xorb hash, hex. }
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        '200':
          description: Uploaded (or already present - see was_inserted).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadXorbResponse' }
        '400':
          description: >-
            Bad prefix/hash, malformed xorb, unsupported/failed
            decompression, or hash mismatch between URL and computed
            content hash.
          content: { text/plain: { schema: { type: string } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '413':
          description: Body exceeds the 128 MiB max xorb size.
          content: { text/plain: { schema: { type: string } } }
        '500':
          description: Storage or internal failure.
          content: { text/plain: { schema: { type: string } } }
    get:
      tags: [CAS]
      summary: Fetch raw xorb bytes
      description: >-
        Requires read scope. Supports a single-range `Range: bytes=<start>-<end>`
        header (both bounds required - no open-ended or multi-range
        requests). Byte-serving endpoint backing the URLs handed out in
        reconstruction fetch_info.
      parameters:
        - { name: prefix, in: path, required: true, schema: { type: string, enum: [default] } }
        - { name: hash, in: path, required: true, schema: { type: string } }
        - name: Range
          in: header
          required: false
          schema: { type: string, example: 'bytes=0-1023' }
      responses:
        '200':
          description: Full xorb bytes (no Range header sent).
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '206':
          description: Partial content for a satisfiable Range request.
          headers:
            Content-Range: { schema: { type: string, example: 'bytes 0-1023/2048' } }
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '400':
          description: Bad prefix/hash.
          content: { text/plain: { schema: { type: string } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: Unknown xorb hash.
        '416':
          description: Malformed or unsatisfiable Range header.
        '500':
          description: Storage read failure.
          content: { text/plain: { schema: { type: string } } }
    head:
      tags: [CAS]
      summary: Fetch xorb size
      description: >-
        Requires read scope. Note (source-verified, undocumented elsewhere):
        unlike GET/POST on this same path, this handler does NOT validate
        {prefix} against "default" - any prefix value is accepted.
      parameters:
        - { name: prefix, in: path, required: true, schema: { type: string } }
        - { name: hash, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Xorb exists.
          headers:
            Content-Length: { schema: { type: integer, format: int64 } }
        '400':
          description: Bad hash.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: Unknown xorb hash.
  /v1/shards:
    post:
      tags: [CAS]
      summary: Upload a serialized shard
      description: >-
        Requires write scope. Body is opaque binary (file-info + xorb-info
        sections, no footer - see internal/shardformat), capped at 16 MiB,
        buffered in full before parsing.
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        '200':
          description: Shard accepted and indexed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadShardResponse' }
        '400':
          description: Malformed shard.
          content: { text/plain: { schema: { type: string } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '413':
          description: Body exceeds the 16 MiB max shard size.
          content: { text/plain: { schema: { type: string } } }
  /v1/reconstructions/{file_id}:
    get:
      tags: [CAS]
      summary: File to xorb/chunk-range reconstruction map (V1)
      description: >-
        Requires read scope. Supports the same single-range Range header
        semantics as xorb fetch, applied against the file's total unpacked
        size.
      parameters:
        - { name: file_id, in: path, required: true, schema: { type: string } }
        - name: Range
          in: header
          required: false
          schema: { type: string, example: 'bytes=0-1023' }
      responses:
        '200':
          description: Reconstruction map for the full file or requested range.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReconstructionResponseV1' }
        '400':
          description: Bad file_id.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: Unknown file_id.
        '416':
          description: Malformed or unsatisfiable Range header.
        '500':
          description: Reconstruction references an unknown xorb, or fetch-URL build failed.
  /v2/reconstructions/{file_id}:
    get:
      tags: [CAS]
      summary: File to xorb/chunk-range reconstruction map (V2, multi-range-optimized)
      description: >-
        Requires read scope. Same underlying data as V1, grouped by xorb
        hash (xorbs) instead of one fetch_info entry per term - fewer
        signed URLs when a file's reconstruction touches the same xorb
        across multiple non-contiguous terms. Same Range/416 semantics as V1.
      parameters:
        - { name: file_id, in: path, required: true, schema: { type: string } }
        - name: Range
          in: header
          required: false
          schema: { type: string, example: 'bytes=0-1023' }
      responses:
        '200':
          description: Reconstruction map for the full file or requested range.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReconstructionResponseV2' }
        '400':
          description: Bad file_id.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: Unknown file_id.
        '416':
          description: Malformed or unsatisfiable Range header.
        '500':
          description: Reconstruction references an unknown xorb, or fetch-URL build failed.
  /v1/chunks/{prefix}/{hash}:
    get:
      tags: [CAS]
      summary: Global chunk-dedup lookup
      description: >-
        Requires read scope. {prefix} must be the literal
        "default-merkledb" (distinct from the xorb-upload prefix). Returns
        the RAW BYTES of whichever previously-uploaded shard referenced
        this chunk hash - not the chunk itself; the client re-parses the
        returned shard to find dedup-eligible chunks. This is the real
        wire contract per xet-core's own OpenAPI spec.
      parameters:
        - { name: prefix, in: path, required: true, schema: { type: string, enum: [default-merkledb] } }
        - { name: hash, in: path, required: true, schema: { type: string }, description: Chunk hash, hex. }
      responses:
        '200':
          description: Raw bytes of a shard that referenced this chunk hash.
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '400':
          description: Bad prefix/hash.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: No uploaded shard has ever referenced this chunk hash.
...